Keep last 10 measurements, not 20.
I feel like I'm throwing noodles at a wall to see what sticks.  The
behavior of MPD just seems so inconsistent.  Some songs seek quickly and
consistently, but others are all over the place.  Sometimes MPD even
gets stuck after seeking and thinks the song is shorter than it actually
is, and it won't seek past that point anymore, so it sounds like a
broken record.  Only solution is to stop and play again, or change songs.