Override more methods in myFloat
This removes the need to cast myFloats back into myFloats whenever doing
math on them.  It is less explicit, though, so I don't know if this is
as in-line with the Zen of Python.  But on the other hand, once you make
a myFloat, it stays a myFloat, so it's less error-prone.  And it looks cleaner.