Don't stop seeking unless range <= 300ms
On some songs, the range might be high (like >500ms), and it could stop
seeking the song if the average difference happened to be <30ms, even if
the actual difference were something like 300ms.  This might fix it, or
it might be better to compare the current difference; we'll see.