Add checks and debugging for weird issue
Sometimes when the playlist changes on the server and has to be changed
on the client, there seems to be an occasional race condition that
causes the song number to be >= the playlist length, which shouldn't
happen.  The code already works around it, so it doesn't do any harm,
but it would be nice to track down the bug someday.  This code may help.